requirement.txt
absl-py==0.11.0
adjustText==0.7.3
amqp==2.6.1
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.3
asgiref==3.3.1
astor==0.8.1
astroid==2.11.2
astunparse==1.6.3
attrs==20.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
bioinfokit==1.0.4
cached-property==1.5.2
cachetools==4.1.1
celery==4.4.2
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cloudpickle==1.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
distlib==0.3.0
Django==3.1.3
django-celery-results==2.0.0
django-extensions==3.1.0
django-filter==2.4.0
django-picklefield==3.0.1
django-rename-app==0.1.2
djangorestframework==3.12.2
drf-yasg==1.20.0
execnet==1.7.1
filelock==3.0.12
flake8==3.8.4
flower==0.9.5
gast==0.3.3
globus-sdk==1.9.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.28.1
grpcio-tools==1.28.1
h5py==2.10.0
httpie==2.3.0
humanize==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==0.18
importlib-resources==1.4.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==5.2.0
ipyparallel==6.2.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==5.10.1
itypes==1.2.0
jedi==0.16.0
Jinja2==2.7.3
joblib==0.14.1
jsonschema==2.4.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.11
lazy-object-proxy==1.7.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
matplotlib-venn==0.11.6
mccabe==0.6.1
mysqlclient==2.0.1
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
openapi-codec==1.3.2
opt-einsum==3.2.1
packaging==20.7
pandas==1.0.4
paramiko==2.7.1
parsl==1.0.0
parso==0.6.2
pathlib2==2.3.5
patsy==0.5.1
pegasus-wms.api==5.0.0
pegasus-wms.common==5.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
platformdirs==2.4.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.8.4
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
pygount==1.2.4
pygraphviz==1.6
PyJWT==1.7.1
pylint==2.13.4
PyMySQL==0.10.1
PyNaCl==1.3.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.2
pytest-celery==0.0.0a1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.1.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.4.1
seaborn==0.11.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.0
sqlparse==0.4.1
statsmodels==0.12.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
textwrap3==0.9.2
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==4.3.3
typed-ast==1.5.2
typeguard==2.7.1
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.2
verta==0.15.9
vine==1.3.0
virtualenv==20.0.18
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0